features(alot): adding product crud(not 'u' i mean update), email notification at every product update and 1 custom exception. PROTECTED ROUTES MAY STILL BE A PROBLEM KEEP IN MIND!!!

+++ src/main/java/com/canureal/cartflame/models/ProductRepository.java
@@ -2,5 +2,5 @@ package com.canureal.cartflame.models;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductRepository extends JpaRepository<Products, Long> {
+public interface ProductRepository extends JpaRepository<Products, Integer> {
 }
